LoyaltyBot/requirements.txt
python-telegram-bot==12
uplink==0.9.4